anh le

+++ lichsu/models.py
@@ -6,10 +6,10 @@ from manhdat.models import *
 # Create your models here.
 class LichSuCamBien(models.Model):
     manhdat = models.ForeignKey(ManhDat, models.SET_NULL, null=True)
-    nhiet_do = models.IntegerField(default=0)
-    do_am = models.IntegerField(default=0)
-    do_am_dat = models.IntegerField(default=0)
-    anh_sang = models.IntegerField(default=0)
+    nhiet_do = models.FloatField(default=0)
+    do_am = models.FloatField(default=0)
+    do_am_dat = models.FloatField(default=0)
+    anh_sang = models.FloatField(default=0)
     created_at = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
@@ -17,10 +17,10 @@ class LichSuCamBien(models.Model):
 
 class LichSuHanhDong(models.Model):
     manhdat = models.ForeignKey(ManhDat, models.SET_NULL, null=True)
-    mai_che = models.IntegerField(default=0)
-    quat_mat = models.IntegerField(default=0)
-    may_tuoi_nuoc = models.IntegerField(default=0)
-    den_chieu_sang = models.IntegerField(default=0)
+    mai_che = models.FloatField(default=0)
+    quat_mat = models.FloatField(default=0)
+    may_tuoi_nuoc = models.FloatField(default=0)
+    den_chieu_sang = models.FloatField(default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     
     def __str__(self):
